Refactoring: split DataView and DataViewer into 2 modules This avoids circular import when designing a DataView subclass in a separate module and adding it to the default views of DataViewer (eg NXdataView)
